requirements.txt
sqlalchemy>=2.0
pandas>=2.0
rich>=13.0
click>=8.0
akshare>=1.10
pyyaml>=6.0
fastapi>=0.110
uvicorn>=0.29
jinja2>=3.1
python-multipart>=0.0.9
Pillow>=10.0
rapidocr_onnxruntime>=1.3
